#1677 Moved Duration and DurationUnit entities to package 'dashboard'.
Improved security tokens for Duration and DurationUnit by reusing ENTITY_TITLE constants.
