+++ platform-dao/src/main/java/ua/com/fielden/platform/dashboard/DurationUnitDao.java
@@ -1,4 +1,4 @@
-package ua.com.fielden.platform.entity;
+package ua.com.fielden.platform.dashboard;
 
 import com.google.inject.Inject;
