+++ platform-dao/src/main/java/ua/com/fielden/platform/ioc/BasicWebServerModule.java
@@ -26,6 +26,10 @@ import ua.com.fielden.platform.dao.IGeneratedEntityController;
 import ua.com.fielden.platform.dao.ISecurityRoleAssociation;
 import ua.com.fielden.platform.dao.IUserAndRoleAssociation;
 import ua.com.fielden.platform.dao.IUserRole;
+import ua.com.fielden.platform.dashboard.DurationCo;
+import ua.com.fielden.platform.dashboard.DurationDao;
+import ua.com.fielden.platform.dashboard.DurationUnitCo;
+import ua.com.fielden.platform.dashboard.DurationUnitDao;
 import ua.com.fielden.platform.domain.metadata.DomainExplorerCo;
 import ua.com.fielden.platform.domain.metadata.DomainExplorerDao;
 import ua.com.fielden.platform.domain.metadata.DomainExplorerInsertionPointCo;
@@ -36,10 +40,6 @@ import ua.com.fielden.platform.domain.metadata.DomainPropertyHolderCo;
 import ua.com.fielden.platform.domain.metadata.DomainPropertyHolderDao;
 import ua.com.fielden.platform.domain.metadata.DomainTypeCo;
 import ua.com.fielden.platform.domain.metadata.DomainTypeDao;
-import ua.com.fielden.platform.entity.DurationCo;
-import ua.com.fielden.platform.entity.DurationDao;
-import ua.com.fielden.platform.entity.DurationUnitCo;
-import ua.com.fielden.platform.entity.DurationUnitDao;
 import ua.com.fielden.platform.entity.EntityDeleteActionDao;
 import ua.com.fielden.platform.entity.EntityEditActionDao;
 import ua.com.fielden.platform.entity.EntityNewActionDao;
@@ -121,19 +121,19 @@ import ua.com.fielden.platform.ui.config.controller.EntityCentreConfigDao;
 import ua.com.fielden.platform.ui.config.controller.EntityLocatorConfigDao;
 import ua.com.fielden.platform.ui.config.controller.EntityMasterConfigDao;
 import ua.com.fielden.platform.ui.config.controller.MainMenuItemDao;
+import ua.com.fielden.platform.web.centre.CentreConfigDeleteActionCo;
 import ua.com.fielden.platform.web.centre.CentreConfigDeleteActionDao;
+import ua.com.fielden.platform.web.centre.CentreConfigDuplicateActionCo;
 import ua.com.fielden.platform.web.centre.CentreConfigDuplicateActionDao;
+import ua.com.fielden.platform.web.centre.CentreConfigNewActionCo;
 import ua.com.fielden.platform.web.centre.CentreConfigNewActionDao;
+import ua.com.fielden.platform.web.centre.CentreConfigShareActionCo;
 import ua.com.fielden.platform.web.centre.CentreConfigShareActionDao;
 import ua.com.fielden.platform.web.centre.CustomisableColumnDao;
-import ua.com.fielden.platform.web.centre.CentreConfigDeleteActionCo;
-import ua.com.fielden.platform.web.centre.CentreConfigDuplicateActionCo;
-import ua.com.fielden.platform.web.centre.CentreConfigNewActionCo;
-import ua.com.fielden.platform.web.centre.CentreConfigShareActionCo;
 import ua.com.fielden.platform.web.centre.ICustomisableColumn;
 import ua.com.fielden.platform.web.centre.LoadableCentreConfigCo;
-import ua.com.fielden.platform.web.centre.OverrideCentreConfigCo;
 import ua.com.fielden.platform.web.centre.LoadableCentreConfigDao;
+import ua.com.fielden.platform.web.centre.OverrideCentreConfigCo;
 import ua.com.fielden.platform.web.centre.OverrideCentreConfigDao;
 import ua.com.fielden.platform.web_api.GraphQLService;
 import ua.com.fielden.platform.web_api.IWebApi;
