+++ platform-pojo-bl/src/main/java/ua/com/fielden/platform/dashboard/DurationUnitCo.java
@@ -1,4 +1,4 @@
-package ua.com.fielden.platform.entity;
+package ua.com.fielden.platform.dashboard;
 
 import ua.com.fielden.platform.dao.IEntityDao;
