+++ platform-pojo-bl/src/main/java/ua/com/fielden/platform/domain/PlatformDomainTypes.java
@@ -7,6 +7,8 @@ import ua.com.fielden.platform.attachment.Attachment;
 import ua.com.fielden.platform.attachment.AttachmentPreviewEntityAction;
 import ua.com.fielden.platform.attachment.AttachmentUploader;
 import ua.com.fielden.platform.attachment.AttachmentsUploadAction;
+import ua.com.fielden.platform.dashboard.Duration;
+import ua.com.fielden.platform.dashboard.DurationUnit;
 import ua.com.fielden.platform.domain.metadata.DomainExplorer;
 import ua.com.fielden.platform.domain.metadata.DomainExplorerInsertionPoint;
 import ua.com.fielden.platform.domain.metadata.DomainProperty;
@@ -14,8 +16,6 @@ import ua.com.fielden.platform.domain.metadata.DomainPropertyTreeEntity;
 import ua.com.fielden.platform.domain.metadata.DomainTreeEntity;
 import ua.com.fielden.platform.domain.metadata.DomainType;
 import ua.com.fielden.platform.entity.AbstractEntity;
-import ua.com.fielden.platform.entity.Duration;
-import ua.com.fielden.platform.entity.DurationUnit;
 import ua.com.fielden.platform.entity.EntityDeleteAction;
 import ua.com.fielden.platform.entity.EntityEditAction;
 import ua.com.fielden.platform.entity.EntityExportAction;
