+++ platform-pojo-bl/src/main/java/ua/com/fielden/platform/ui/menu/sample/MiDuration.java
@@ -1,6 +1,6 @@
 package ua.com.fielden.platform.ui.menu.sample;
 
-import ua.com.fielden.platform.entity.Duration;
+import ua.com.fielden.platform.dashboard.Duration;
 import ua.com.fielden.platform.entity.annotation.EntityType;
 import ua.com.fielden.platform.ui.menu.MiWithConfigurationSupport;
