+++ platform-web-ui/src/main/java/ua/com/fielden/platform/web/centre/CentreUpdater.java
@@ -90,11 +90,11 @@ import java.util.stream.Stream;
 import org.apache.log4j.Logger;
 
 import ua.com.fielden.platform.dao.IEntityDao;
+import ua.com.fielden.platform.dashboard.Duration;
 import ua.com.fielden.platform.domaintree.IDomainTreeEnhancerCache;
 import ua.com.fielden.platform.domaintree.centre.ICentreDomainTreeManager.ICentreDomainTreeManagerAndEnhancer;
 import ua.com.fielden.platform.domaintree.centre.IOrderingRepresentation.Ordering;
 import ua.com.fielden.platform.entity.AbstractEntity;
-import ua.com.fielden.platform.entity.Duration;
 import ua.com.fielden.platform.entity.factory.ICompanionObjectFinder;
 import ua.com.fielden.platform.entity.query.DbVersion;
 import ua.com.fielden.platform.entity.query.fluent.EntityQueryProgressiveInterfaces.ICompoundCondition0;
